// tests/e2e/orm/schemas/custom-type-primitive/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "postgresql"
    } as const;
    models = {
        User: {
            name: "User",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("nanoid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("nanoid") as FieldDefault
                },
                name: {
                    name: "name",
                    type: "UserName",
                    attributes: [{ name: "@length", args: [{ name: "min", value: ExpressionUtils.literal(2) }, { name: "max", value: ExpressionUtils.literal(16) }] }] as readonly AttributeApplication[]
                },
                contacts: {
                    name: "contacts",
                    type: "Contact",
                    array: true,
                    attributes: [{ name: "@phone" }] as readonly AttributeApplication[]
                },
                age: {
                    name: "age",
                    type: "Age",
                    optional: true,
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.literal(18) }] }, { name: "@gte", args: [{ name: "value", value: ExpressionUtils.literal(18) }] }] as readonly AttributeApplication[],
                    default: "18" as FieldDefault
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        }
    } as const;
    typeDefs = {
        UserName: {
            name: "UserName",
            base: "String",
            fields: {
                this: {
                    name: "this",
                    type: "String",
                    attributes: [{ name: "@length", args: [{ name: "min", value: ExpressionUtils.literal(2) }, { name: "max", value: ExpressionUtils.literal(16) }] }] as readonly AttributeApplication[]
                }
            }
        },
        Contact: {
            name: "Contact",
            base: "String",
            fields: {
                this: {
                    name: "this",
                    type: "String",
                    attributes: [{ name: "@phone" }] as readonly AttributeApplication[]
                }
            }
        },
        Age: {
            name: "Age",
            base: "Int",
            fields: {
                this: {
                    name: "this",
                    type: "Int",
                    attributes: [{ name: "@gte", args: [{ name: "value", value: ExpressionUtils.literal(18) }] }] as readonly AttributeApplication[]
                }
            }
        }
    } as const;
    authType = "User" as const;
    plugins = {};
}
export const schema = new SchemaType();